[libc] Refactor pwd database state into passwd namespace
Moved password database file state (pwd_file, line_buffer, pwd_entry)
and reading/rewinding/closing logic into namespace passwd in pwd_utils.h / pwd_utils.cpp

* Decoupled setpwent and endpwent from depending on getpwent in CMake
* Added target guards for File and platform_file in src/pwd and test/src/pwd CMakeLists.txt
* Reused open() in read_next() and simplified read_line return
* Updated getpwent_test to use ErrnoCheckingTest, ASSERT_ERRNO_EQ,
  internal::strlcpy, and passwd::TESTONLY_set_passwd_path
