+++ libc/src/pwd/CMakeLists.txt
@@ -1,3 +1,7 @@
+if(NOT TARGET libc.src.__support.File.file OR NOT TARGET libc.src.__support.File.platform_file)
+  return()
+endif()
+
 add_entrypoint_object(
   getpwent
   SRCS
@@ -9,9 +13,6 @@ add_entrypoint_object(
     libc.src.errno.errno
     libc.src.__support.common
     libc.src.__support.macros.config
-    libc.src.__support.File.file
-    libc.src.__support.File.platform_file
-    libc.hdr.stdio_macros
     .pwd_utils
 )
 
@@ -25,7 +26,7 @@ add_entrypoint_object(
     libc.src.errno.errno
     libc.src.__support.common
     libc.src.__support.macros.config
-    .getpwent
+    .pwd_utils
 )
 
 add_entrypoint_object(
@@ -38,7 +39,7 @@ add_entrypoint_object(
     libc.src.errno.errno
     libc.src.__support.common
     libc.src.__support.macros.config
-    .getpwent
+    .pwd_utils
 )
 
 add_object_library(
@@ -49,8 +50,12 @@ add_object_library(
     pwd_utils.cpp
   DEPENDS
     libc.hdr.errno_macros
+    libc.hdr.stdio_macros
     libc.hdr.types.struct_passwd
     libc.src.string.string_utils
+    libc.src.__support.CPP.span
+    libc.src.__support.File.file
+    libc.src.__support.File.platform_file
     libc.src.__support.ctype_utils
     libc.src.__support.error_or
     libc.src.__support.str_to_integer
