+++ libc/src/pwd/pwd_utils.h
@@ -25,6 +25,22 @@ namespace internal {
 ErrorOr<struct passwd> parse_passwd_line(char *line);
 
 } // namespace internal
+
+namespace passwd {
+
+// Overrides the default password file path for testing purposes.
+void TESTONLY_set_passwd_path(const char *path);
+
+// Opens or rewinds the password file.
+ErrorOr<int> open();
+
+// Closes the password file.
+ErrorOr<int> close();
+
+// Reads the next entry from the password database.
+ErrorOr<struct passwd *> read_next();
+
+} // namespace passwd
 } // namespace LIBC_NAMESPACE_DECL
 
 #endif // LLVM_LIBC_SRC_PWD_PWD_UTILS_H
