+++ libc/src/pwd/setpwent.cpp
@@ -14,14 +14,12 @@
 #include "src/pwd/setpwent.h"
 #include "src/__support/common.h"
 #include "src/__support/libc_errno.h"
-#include "src/pwd/getpwent.h"
+#include "src/pwd/pwd_utils.h"
 
 namespace LIBC_NAMESPACE_DECL {
 
 LLVM_LIBC_FUNCTION(void, setpwent, ()) {
-  // setpwent_impl opens or rewinds the password file. If an error occurs,
-  // it returns an Error with an errno value which is set here.
-  auto res = setpwent_impl();
+  auto res = passwd::open();
   if (!res.has_value())
     libc_errno = res.error();
 }
